pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sim-ui"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pywebview"]

[tool.hatch.build.targets.wheel]
packages = ["sim_ui"]

[tool.hatch.build.targets.wheel.force-include]
"sim_ui/dist" = "sim_ui/dist"
